+++ app/Views/categorias.view.php
@@ -10,22 +10,27 @@
     ?>
     <div class="col-12">
         <div class="card shadow mb-4">
-            <div
-                class="card-header py-3 d-flex flex-row align-items-center justify-content-between">
+            <div class="card-header py-3 d-flex flex-row align-items-center justify-content-between">
                 <div class="col-6">
-                <h6 class="m-0 installfont-weight-bold text-primary">Proveedores</h6> 
+                    <h6 class="m-0 installfont-weight-bold text-primary">Proveedores</h6> 
                 </div>
-                <div class="col-6">                       
-                    <div class="m-0 font-weight-bold justify-content-end">
-                        <a href="/categorias/add/" class="btn btn-primary ml-1 float-right"> Nueva Categoría <i class="fas fa-plus-circle"></i></a>
-                    </div>                    
-                </div>
+                <?php
+                if (strpos($_SESSION['permisos']['usuarios_sistema'], 'w') !== false) {
+                    ?>
+                    <div class="col-6">                       
+                        <div class="m-0 font-weight-bold justify-content-end">
+                            <a href="/categorias/add/" class="btn btn-primary ml-1 float-right"> Nueva Categoría <i class="fas fa-plus-circle"></i></a>
+                        </div>                    
+                    </div>
+                    <?php
+                }
+                ?>
             </div>
             <!-- Card Body -->
             <div class="card-body" id="card_table">
                 <div id="button_container" class="mb-3"></div>
                 <?php
-                if (count($categorias) > 0) {                    
+                if (count($categorias) > 0) {
                     ?>
                     <!--<form action="./?sec=formulario" method="post">                   -->
                     <table id="tabladatos" class="table table-striped">                    
@@ -40,16 +45,25 @@
                             <?php
                             foreach ($categorias as $c) {
                                 ?>
-                                <tr class="<?php #echo $p['pais'] != 'España' ? 'table-warning' :  '';  ?>">
+                                <tr class="<?php #echo $p['pais'] != 'España' ? 'table-warning' :  '';      ?>">
                                     <td><?php echo $c['id_categoria']; ?></td>
                                     <td><?php echo $c['nombre_categoria']; ?></td>                                     
                                     <td><?php echo $c['fullName']; ?></td>   
                                     <td> 
                                         <a href="/categorias/view/<?php echo $c['id_categoria']; ?>" class="btn btn-default ml-1"><i class="fas fa-eye"></i></a>
-                                        <a href="/categorias/edit/<?php echo $c['id_categoria']; ?>" class="btn btn-success ml-1"><i class="fas fa-edit"></i></a>
-                                        <a href="/categorias/delete/<?php echo $c['id_categoria']; ?>" class="btn btn-danger ml-1"><i class="fas fa-trash"></i></a>
+                                        <?php
+                                        if (strpos($_SESSION['permisos']['usuarios_sistema'], 'w') !== false) {
+                                            ?>
+                                            <a href="/categorias/edit/<?php echo $c['id_categoria']; ?>" class="btn btn-success ml-1"><i class="fas fa-edit"></i></a>
+                                            <?php
+                                        }
+                                        if (strpos($_SESSION['permisos']['usuarios_sistema'], 'd') !== false) {
+                                            ?>
+                                            <a href="/categorias/delete/<?php echo $c['id_categoria']; ?>" class="btn btn-danger ml-1"><i class="fas fa-trash"></i></a>
+                                            <?php
+                                        }
+                                        ?>
                                     </td>
-
                                 </tr>
                                 <?php
                             }
